fix: move view key page dialog to the correct folder
