+++ src/components/app/userActionFormShareOnTwitter/common/constants.ts
@@ -4,5 +4,3 @@ export enum SectionNames {
 }
 
 export const ANALYTICS_NAME_USER_ACTION_FORM_SHARE_ON_TWITTER = 'User Action Form Share on Twitter'
-
-export const ANALYTICS_NAME_USER_ACTION_VEW_KEY_PAGE = 'User Action Form View Key Page'
