+++ src/components/app/userActionFormViewKeyPage/constants.ts
@@ -1,0 +1,1 @@
+export const ANALYTICS_NAME_USER_ACTION_VEW_KEY_PAGE = 'User Action Form View Key Page'
